Setup ESLint and TravisCI. Reorganized the ReactApp. Start using JSX and SCSS.
